offline-version/requirements.txt
streamlit
openai-whisper
torch
pyttsx3
numpy
sounddevice
scipy
